condaenv.sesmubs6.requirements.txt
google-generativeai>=0.7.2